Day 4 slot C sub-slice 4C.d.iv: LankaEvents.Application Money/Currency operator-form migration
Consult #13 Q2 pattern (previously applied to LankaEvents.Domain) extended
to LankaEvents.Application caller sites.

Transitional shim: MoneyBuilder.cs added at
src/Products/LankaEvents/LankaEvents.Application/Common/. Exposes
Result<Money>-shaped Create() + TryParseCurrency() so caller signatures
stay identical during the intermediate cutover. Delete when LankaEvents
handlers rewrite fully to operator form (Day 5 slot B / Wave 6.5.f).

Sweeps applied:
- `Money.Create(a, c)` -> `MoneyBuilder.Create(a, c)` across all LankaEvents
  .Application/*.cs (28 files, ~90 sites).
- `SomeXxxCurrency.HasValue` -> `SomeXxxCurrency != null` (currency-typed
  nullable reference type; nullable value type pattern no longer applies
  since SharedKernel.Money.Currency is a class).
- `SomeXxxCurrency.Value` -> `SomeXxxCurrency` (Currency? unwrap).
- `Enum.TryParse<Currency>(...)` -> `MoneyBuilder.TryParseCurrency(...)`
  (Currency is no longer an enum; wraps Currency.TryFromCode Maybe<>).
- `Enum.Parse<Currency>(...)` -> `Currency.FromCode(...)` (1 site).
- `SharedKernel.Money` using added to 28 files.

Non-shim fixes:
- Address.Create + GeoCoordinate.Create static factories added to
  LankaEvents.Domain/ValueObjects/ContactInfoPrimitives.cs (matches caller
  Result<T>-shaped expectations; decimal overload for GeoCoordinate).
- `BB.Domain.Shared.ValueObjects.Email` fully-qualified refs in
  EventCancellationEmailJob + EventNotificationEmailJob rewritten to
  LankaEvents.Domain.ValueObjects.Email (Email stub landed there in 4C.a).
- EventMappingProfile.cs: added (decimal) cast on GeoCoordinate.Latitude/
  Longitude MapFrom lambdas (Coordinates.Latitude is now double per stub;
  DTO destination is decimal?).
- GetEventsQueryHandler.cs: same (decimal) cast at CalculateDistance call
  sites (2 places).
- GetEventByIdQueryHandler.cs: EmailGroupSummaryDto ctor now positional
  (Contracts version has 8-arg record ctor; supplied null/Guid.Empty/0
  placeholders for fields not accessible from the LankaEvents side pending
  a proper cross-module query pattern).
- 3 DTOs (GroupPricingTierDto, RevenueBreakdownDto, TicketTierDto) got
  `= null!` on their `Currency` properties (CS8618 non-nullable-init
  warning; property is set by mapping profile at runtime).
- EmailDateTimeHelper aliased to Contracts version in event handlers.
- MoneyBuilder.Create signature accepts Currency? and returns
  Result<Money>.Failure when null (CS8604 nullable warning fix).

Build state trajectory:
- Pre-4C.d.iv: 218 errors (LankaEvents.Application 184 dominated).
- Post-4C.d.iv: 140 errors. LankaEvents.Application 184 -> 0 ✅.
  Unmasked downstream: Payments.Application 34 (same Money/Currency shape;
  handled in Day 5 slot C / Wave 6.5.g), LankaConnect.Infrastructure 12.
  Media.Api 132 unchanged (baseline package/using-refs; Day 5 slot A).

T-triggers: T6 (MoneyBuilder shim + ContactInfoPrimitives new factory
methods).
S-class: n/a (LankaEvents.Application Command/Query handler rewrites; no
runtime path touched by shim addition).

Follow-up: MoneyBuilder shim deletion + rewrite to operator form (Consult
#13 Q2 canonical form) is Day 5 slot B territory (Wave 6.5.f LankaEvents
handler migration). Track as [[followup-track]] under MoneyBuilder shim
retirement.

+++ src/Products/LankaEvents/LankaEvents.Application/Queries/GetEvents/GetEventsQueryHandler.cs
@@ -578,8 +578,8 @@ public class GetEventsQueryHandler : IQueryHandler<GetEventsQuery, IReadOnlyList
                         var distance = CalculateDistance(
                             metroData.Value.Latitude,
                             metroData.Value.Longitude,
-                            e.Location!.Coordinates!.Latitude,
-                            e.Location.Coordinates.Longitude);
+                            (decimal)e.Location!.Coordinates!.Latitude,
+                            (decimal)e.Location.Coordinates.Longitude);
                         // Convert radius from miles to kilometers (1 mile = 1.60934 km)
                         var radiusKm = metroData.Value.RadiusMiles * 1.60934;
                         return distance <= radiusKm;
@@ -635,8 +635,8 @@ public class GetEventsQueryHandler : IQueryHandler<GetEventsQuery, IReadOnlyList
                 var distance = CalculateDistance(
                     latitude,
                     longitude,
-                    e.Location.Coordinates.Latitude,
-                    e.Location.Coordinates.Longitude);
+                    (decimal)e.Location.Coordinates.Latitude,
+                    (decimal)e.Location.Coordinates.Longitude);
                 eventsWithCoords.Add((e, distance));
             }
             else
